ISSUE #2756 fix swagger, move revision middleware into the right place
